requirements.txt
django~=5.0.7
djangorestframework~=3.15.2
psycopg2-binary
Pillow
django-filter
django-cors-headers
requests
django-extensions
channels_redis
daphne
channels
django-extensions
google-cloud-vision
celery
redis
segno
weasyprint
django-celery-beat